Change method of getting server IP
Using the PHP $_SERVER['SERVER_ADDR'] variable to get the server IP can be inaccurate. If, for example, the servers PHP (php-fpm) is configured with Apache listening on 127.0.0.1 behind Nginx on the public IPs then the reported IP is not correct. Doing this DNS lookup at least gives a better chance of getting a correct IP.

I'm considering adding a menu allowing users to specify a header to look for; then they, or their host, can setup a Header passing config from nginx to allow Apache (and PHP) for the plugin to find.